pydataflow/pydataflow/requirement.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coverage==4.5.1
croniter==0.3.25
cryptography==2.3.1
cx-Oracle==7.0.0
Cython==0.29.7
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.1
django-allauth==0.37.1
django-crispy-forms==1.7.2
django-datatable==0.3.1
django-extensions==2.1.2
django-filter==2.2.0
django-forms-bootstrap==3.1.0
django-import-export==1.1.0
django-multiselectfield==0.1.8
django-tables2==2.0.6
django-widget-tweaks==1.4.1
django-widgets-improved==1.5.0
djangorestframework==3.9.3
djangorestframework-datatables==0.5.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.8
idna==2.7
isort==4.3.21
jdcal==1.4
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
mysql-connector-python==8.0.11
mysqlclient==1.3.13
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.9
protobuf==3.7.1
pycodestyle==2.5.0
pycparser==2.18
pyflakes==2.1.1
pylint==2.3.1
pymssql==2.1.4
PyMySQL==0.9.2
pyodbc==4.0.26
python-crontab==2.3.5
python-dateutil==2.7.4
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
tablib==0.12.1
typed-ast==1.4.0
unicodecsv==0.14.1
urllib3==1.23
wrapt==1.11.2
xlrd==1.1.0
xlwt==1.3.0
